refactor: remove BidirectionalPathService and move to PathService
