+++ src/main/java/nextstep/subway/maps/map/application/FareMapService.java
@@ -31,10 +31,10 @@ public class FareMapService {
 
     private final LineService lineService;
     private final StationService stationService;
-    private final BidirectionalPathService pathService;
+    private final PathService pathService;
     private final DiscountFareCalculator fareCalculator;
 
-    public FareMapService(LineService lineService, StationService stationService, BidirectionalPathService pathService,
+    public FareMapService(LineService lineService, StationService stationService, PathService pathService,
         DiscountFareCalculator fareCalculator) {
         this.lineService = lineService;
         this.stationService = stationService;
