+++ src/test/java/nextstep/subway/maps/map/application/FareMapServiceTest.java
@@ -42,7 +42,7 @@ public class FareMapServiceTest {
     @Mock
     private StationService stationService;
     @Mock
-    private BidirectionalPathService pathService;
+    private PathService pathService;
     @Mock
     private DiscountFareCalculator fareCalculator;
